Integrate next-auth with prisma ORM
