+++ package.json
@@ -21,7 +21,11 @@
     "@ai-sdk/azure": "^1.0.7",
     "@googlemaps/react-wrapper": "^1.1.35",
     "@googlemaps/typescript-guards": "^2.0.1",
+    "@neondatabase/serverless": "^1.0.1",
+    "@next-auth/prisma-adapter": "^1.0.7",
     "@playwright/test": "^1.49.0",
+    "@prisma/adapter-neon": "^6.9.0",
+    "@prisma/client": "^6.9.0",
     "@radix-ui/react-dialog": "^1.1.4",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/themes": "^1.1.0",
@@ -43,7 +47,7 @@
     "marked": "^4.0.18",
     "motion": "^11.13.3",
     "next": "^15.0.3",
-    "next-auth": "^4.24.10",
+    "next-auth": "^4.24.11",
     "next-pwa": "^5.6.0",
     "next-react-svg": "^1.1.3",
     "next-seo": "^5.5.0",
@@ -55,7 +59,8 @@
     "react-simple-typewriter": "^3.0.1",
     "swiper": "^8.3.1",
     "tailwind-merge": "^2.3.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "ws": "^8.18.2"
   },
   "devDependencies": {
     "@fullhuman/postcss-purgecss": "^4.1.3",
@@ -74,6 +79,7 @@
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/react-portal": "^4.0.4",
+    "@types/ws": "^8.18.1",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
     "autoprefixer": "^10.4.4",
@@ -94,10 +100,12 @@
     "postcss": "^8.4.12",
     "prettier": "^3.4.2",
     "prettier-plugin-tailwindcss": "^0.6.9",
+    "prisma": "^6.9.0",
     "rimraf": "^3.0.2",
     "tailwindcss": "^3.0.23",
     "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
+    "tsx": "^4.19.4",
     "typescript": "^5.6.3"
   },
   "lint-staged": {
