+++ src/lib/prisma.ts
@@ -1,0 +1,15 @@
+import { PrismaClient } from "@prisma/client";
+import { PrismaNeon } from "@prisma/adapter-neon";
+import { neonConfig } from "@neondatabase/serverless";
+import ws from "ws";
+
+neonConfig.webSocketConstructor = ws;
+
+const globalForPrisma = global as unknown as { prisma: PrismaClient };
+const connectionString = `${process.env.DATABASE_URL}`;
+const adapter = new PrismaNeon({ connectionString });
+
+const prisma = globalForPrisma.prisma || new PrismaClient({ adapter });
+if (process.env.NODE_ENV === "development") globalForPrisma.prisma = prisma;
+
+export default prisma;
